LayoutTests/platform/ios-simulator/editing/selection/caret-bidi-first-and-last-letters-expected.txt

CONSOLE MESSAGE: line 42: IndexSizeError: DOM Exception 1: Index or size was negative, or greater than the allowed value.
This tests that clicking in a contenteditable div will set the caret in the right position for LTR/RTL text in a RTL/LTR block.

Clicking the first letter of RTL text in a LTR block
שדה בור
Clicking the last letter of RTL text in a LTR block
שדה בור
Clicking the first letter of LTR text in a RTL block
WebKitW
Clicking the last letter of LTR text in a RTL block
WebKitW
